Added check for non-malloced objects in RCD
Summary:
For an object that is not created using malloced/realloced, running RCD on it is pointless. Hence adding a condition to check if object we are considering for RCD is malloced or not. Apparently for some classes in iOS 14 (libraries) the objects are not malloc based.
The classes altered in iOS 14 underneath might be putting the nodes in segments other than heap. Unaware of what all these classes are safest choice is to check if the object is malloced at the first placed, if not ignore it.

Examples:
`__NSGlobalBlock__`
(integer constants)
(string constants)
Empty arrays `__NSArray0`

Fellow developer on RCD open source, already put up a similar PR for this.
https://fburl.com/a78ug38s
Probably merge from there (to encourage it more), unsure about the protocol here.

Differential Revision: D23822041

fbshipit-source-id: 1087030d379590774b1a4db9addf89f5d3a269f8