Altered test cases to make sure that blocks are always malloc-ed
Summary:
Alter the test code for `FBRetainCycleDetectorTests.mm`

Apparently the different version of LLVM we run on *Sandcastle (Jackalope tier use an opensource LLVM compiler)* optimizes the code in a way that some block objects become stack based (`__NSStackBlock__`).

While running it on a mac will not fail because the version of LLVM on Mac (Apple's version) does not do that optimization.

Hence we fail our test.

**Solution:**
Sending a `- (id)copy;` message to the block object, making sure that it is heap (`__NSMallocBlock__`) based.

Reviewed By: youngdfb, narissiam

Differential Revision: D24067445

fbshipit-source-id: 5207d7f17897c62836394382753cd4f1e7d0f56a