docs: add GH button in support of Ukraine (#128)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebook/FBRetainCycleDetector/pull/128

Reviewed By: sarora5

Differential Revision: D34628567

Pulled By: dmitryvinn-fb

fbshipit-source-id: 20ae973e6ee8d5fed935aeb2e0f166cdf7ab1002