Take the layout out of FBObjectReference
Summary:
I'm using D53445283 as a starting point to support Swift objects in RCD.
This diff is taking out of FBObjectReference the layout because for Swift it is not relevant. We will end up with two kinds of references: References with layout and references for Swift.

There is no change in functionality.

 {F1496115031}

Reviewed By: thegreatwallfb

Differential Revision: D56569093

fbshipit-source-id: 0a0cea4b255164b1c10e7528432db913ac53444a