Adding FBSwiftReference to detect swift objects.
Summary:
To detect cycles with swift object we need to check if and object is a SwiftClass or not. If is a SwiftClass we use `FBGetStrongReferencesForSwiftClass` to get all the references. The reference of a SwiftClass  are retrieved using SwiftIntrospector.swift, this code is [Unlicense license](https://github.com/gor-gyolchanyan-swift/introspection-kit?tab=Unlicense-1-ov-file) and we can change it and include in our codebase without concerns.

This detect cycles that include:
- Objc-backed Swift classes
- Pure Swift classes

To enable detection of SwiftClass we need to set `shouldIncludeSwiftObjects` to true in the configuration.

Reviewed By: thegreatwallfb

Differential Revision: D56762443

fbshipit-source-id: 345f6c6e730e8b451803a74b77bd4fd4612227b4