Fix Crash for KVO classes
Summary:
There is a difference in the types supported by Objc and Swift. When a class implements Key-Value Observing ([KVO](https://developer.apple.com/library/archive/documentation/Cocoa/Conceptual/KeyValueObserving/KeyValueObserving.html))  there is a  dynamically-generated subclass created. This classes are not supported by Swift type:link(https://forums.swift.org/t/type-of-vs-object-getclass-difference/59404/2)]
and made Swift Introspection crash.

To fix this we need to get the class in swift calling:
  let typeObj = type(of: object)
And avoid using in swift the class we got in Objc using:
  object_getClass()

Reviewed By: thegreatwallfb

Differential Revision: D58285366

fbshipit-source-id: 8c453eb2d681843d532d473efc9e96443581fb52