Fix NSProxy crash when RCD is running
Summary:
We've been seeing a stack overflow crash due to NSProxy methods infinitely recursing. Using `isProxy` is the safe way to check if an object is an NSProxy so we can avoid `respondsToSelector` which will cause the recursion.

Something I learned: if we use `isKindOfClass: NSProxy` to check if the object is a proxy, that will also cause recursion. That is why `isProxy` is used here.

I'm removing the `methodSignatureForSelector` and `forwardInvocation` overrides which were the initial fix for an NSProxy crash (unimplemented override) but those methods caused recursion which crashed the SMB app, so it's probably best to remove them and avoid similar crashes in other places.

Differential Revision: D77393508

fbshipit-source-id: 8adbc7cc724e7c67113f4212e21d0cc6e755eaca