Fix block introspection for blocks with extended layouts
Summary:
The block introspection was still breaking for a somewhat rare case where a block was represented with the extended layout. This primarily happens either when the reference count of a specific kind (i.e. strong, weak, __block) can't fit in 4 bits, or if it captures anything other than those three (e.g. non-objects).

When this happens, the block storage (i.e. the chunk of memory after the block literal struct) looks different than it does for the compact layouts in the sense that strong objects can be interleaved with __block objects, rather than being listed in the oder of strong, __block, weak. It can also interleave non-objects in there, which is fun.

This means that the logic we used before—which was the same for compact and extended layouts—would fail whenever it found these interleaved types.

To get around that this diff separates the logic for each block layout, and updates the extended block layout's logic to collect the objects as it's iterating through the layout, so that it can increment the word offset even for captured objects and values we don't care about.

Reviewed By: twhitt14

Differential Revision: D77424649

fbshipit-source-id: cbff553b75b47255be1be0d66bad7bad0231b77f